Redo the DKMS debian packaging
Rework this to follow the standard debian way of making dkms packages.
The postint and prerm scripts can go away and the standard debian
package dkms code will work.

Instead of ignoring the dts file in the source code and installing a
dtbo binary, actually build the dts source.

This package will now upgrade and uninstall property.  It was leaving
things behind before and so it didn't entirely work.

Architecture should be "all" not "any".

It still depends on raspberrypi-kernel, even though it could work with
almost any kernel.  It doesn't need to run raspbian or even be ARM.