Rewrite matrixio mic audio driver
This driver was doing a lot of things wrong.  Most noticeable, it
depending on the pointer callback to reset the audio buffer back to the
beginning.  If this wasn't called soon enough, like in an audio overrun,
it would just keep writing past the end of the buffer and clobber the
kernel.

It also did not handle overruns correctly because it didn't implement a
trigger method that would stop the audio process on an xrun.

The copy to user code was very inefficient.  Rework this so it uses a
standard alsa vmalloc buffer, which can be mmaped.  When data from the
SPI transfer is received from the FPGA, interleave it the process of
writing it into this buffer.  It would be more correct to make this
device provide non-interleaved data to ALSA, but most software wouldn't
work with it then.

Use one single SPI xfer to read all channels.  This is much more
efficient than one xfer per channel.

Use atomic and spinlocks.  Certain alsa callbacks, like trigger and
pointer, are called from atomic context and don't work with a mutex.
The atomic avoids some locking.

The matrixio generates constant IRQs and there is no way to stop it.  To
avoid background audio processing costs, use a flag to get the interrupt
handler to short circuit and not perform SPI xfers and audio interleaving
when no audio stream is in progress.  This also fixes races with the ALSA
core code vs the interrupt handler and audio work queue.

Move the work function to a high priority work queue.  At 48 kHz, it has
only 10.7 ms to finish the SPI xfer of the data before the FPGA will
buffer swap.  It won't always make this without being high priority.  I
think a threaded IRQ handler might be even better, but it's still a work
queue for now.  Also use a flag to detect most instances of the FPGA
buffer swapping before the SPI xfer is able to finish getting the
buffer.  Nothing can be done, but at least it's possible to tell why the
audio data is not correct.

Split the matrixio_substream struct into two versions, one for mic and
one for playback, since they are now very different.  The playback
struct is cleaned up a bit to remove unused fields but the playback
driver is largely left alone.

There is still the single global matrixio_substream pointer, which is
not the right design for a driver.  I think maybe this shouldn't be an
asoc driver and that would make that easier to fix by giving access to
the runtime private to store the pointer.

Redo the hardware params.  It does support MMAP now.  It never supported
pause.  The period size is fixed from the FPGA at 512 frames.  Add
constraints to enforce this.  Allow for a larger buffer, 256kB, to allow
for reasonable userspace latency at high channel counts and sample
races.

The sample rates reported as supported to ALSA were not the same as
actual rates supported with CIC and FIR filter values.  Fix this.  Also
avoid cases where a non-supported sample rate would silently work, but
use the wrong decimation, filter, etc. values from whatever rate had
been programmed last.

Note that there are still filter values for a few non-standard rates.  I
didn't add support using a knot constaint for them.