Fix some issues with the Makefile
Split the Makefile out into a Kbuild plus a Makefile.  Works better with
dkms.

Use $(MAKE) so jobserver works.

Allow setting KDIR to use different kernel than the running one, which
allows one to build on an x86 workstation with cross compiler.

Use CURDIR since PWD is the latter is shell variable that is not always
defined

Don't call sudo in the makefile.  Does not play nice with debian package
building.  Use "sudo make install" if you want to install as root.