Rewrite the core SPI transfer functions
These were inefficient and also imposed a 4kB - 2 max size on the data
in one operation.

This version uses two methods, based on the size of the data.

For "small" data, a single transfer is done with the combined address
and data in the transfer buffer.  The data must be copied in/out, but
this this is faster than the overhead of two transfers.

For "large" data, two transfer are used (in one atomic spi message), so
that the data does not need to be copied.  This avoids needing
arbitrarily large bounce buffers.

Get rid of the spi_lock spinlock.  It was useless.

There's no need to save the default spi bus speed and set in each
transfer.  They will automatically use the default bus speed if the
value is left as zero.

The reg read function can be made static.  The reg write function is
still used in other modules, which should switch to regmap and also not
hold spinlocks while they execute spi transactions.